Add BitBoxApp connection option
The BitBoxApp used by the BitBox hardware wallets uses a regular
Electrum backend and supports both clearnet and Tor, using TLS/SSL or
plaintext connections.

For the BitBoxApp, we will consider adding a QR code scan feature in the
server input field, also warning the user if the Tor proxy needs to be
enabled first.